Fix punctuation
Making a bit of an executive call here: It's better to be consistent and remove punctuation across a section if there's even one incomplete sentence. If every bullet within a section is a complete sentence, then it's ok to keep bullets. Consistent formatting > grammatical consistency for readers imo, differently punctuated bullets in the same list can be jarring